requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
google-generativeai>=0.8.0
qdrant-client>=1.9.0
fastapi>=0.110.0
uvicorn>=0.27.0
